requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
sqlalchemy
alembic
psycopg2-binary
celery[redis]
redis
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-celery
opentelemetry-instrumentation-sqlalchemy
langchain
langchain-openai
langchain-community
langchain-text-splitters
openai
faiss-cpu
pymilvus
pypdf
docx2txt
pytest
httpx
